changed health calculation
Health didn't work like I thought it did, so I changed it to how it is
suppose to calculate. Now the program asks for all of your stats, as
they are all needed to determine your health. The calculation is still
slightly off though, still not sure why. Added some more info in the
README