Added GUI Interface
Made a GUI Interface for the program using tkinder. This is my first
time building a GUI, so the interface is a little strange but it works.
This does not include the health total, since it would be hard to add
and there is no point until I fix the calculation. There is one bug,
which will be posted as a bug under the github page.